Refactor HolidayTableComponent into container and table
Split HolidayTableComponent into HolidayTableContainer for data fetching and HolidayTable for presentation. This improves separation of concerns and code maintainability.
